OSIS-833 Check requirement/allocation entity must be same for specific container types
