+++ base/forms/learning_unit_create.py
@@ -37,7 +37,8 @@ from base.forms.bootstrap import BootstrapForm
 from base.forms.utils.choice_field import add_blank
 from base.models.campus import find_main_campuses, Campus
 from base.models.entity_version import find_main_entities_version
-from base.models.enums.learning_container_year_types import LEARNING_CONTAINER_YEAR_TYPES, INTERNSHIP
+from base.models.enums.learning_container_year_types import LEARNING_CONTAINER_YEAR_TYPES, INTERNSHIP, \
+    LEARNING_CONTAINER_YEAR_TYPES_MUST_HAVE_SAME_ENTITIES
 from base.models.enums.learning_container_year_types import LEARNING_CONTAINER_YEAR_TYPES_FOR_FACULTY
 from base.models.enums.learning_unit_management_sites import LearningUnitManagementSite
 from base.models.enums.learning_unit_periodicity import PERIODICITY_TYPES
@@ -165,14 +166,21 @@ class LearningUnitYearForm(BootstrapForm):
 
     def clean(self):
         cleaned_data = super().clean()
-        self._check_if_acronym_already_exists(cleaned_data)
+        if self.errors:
+            return cleaned_data
 
+        self._check_if_acronym_already_exists(cleaned_data)
         if 'internship_subtype' in self.fields \
                 and cleaned_data.get("container_type") == INTERNSHIP \
                 and not (cleaned_data['internship_subtype']):
             self.add_error('internship_subtype', _('field_is_required'))
         if not cleaned_data["common_title"] and not cleaned_data["specific_title"]:
             self.add_error("common_title", _("must_set_common_title_or_specific_title"))
+
+        requirement_entity = cleaned_data["requirement_entity"]
+        allocation_entity = cleaned_data["allocation_entity"]
+        container_type = cleaned_data["container_type"]
+        self._are_requirement_and_allocation_entities_valid(requirement_entity, allocation_entity, container_type)
         return cleaned_data
 
     def _check_if_acronym_already_exists(self, cleaned_data):
@@ -182,6 +190,11 @@ class LearningUnitYearForm(BootstrapForm):
             learning_unit_years_list = self._get_existing_acronym_list(academic_year, acronym)
             if acronym in learning_unit_years_list:
                 self.add_error('acronym', _('already_existing_acronym'))
+
+    def _are_requirement_and_allocation_entities_valid(self, requirement_entity, allocation_entity, container_type):
+        if requirement_entity != allocation_entity and \
+                container_type in LEARNING_CONTAINER_YEAR_TYPES_MUST_HAVE_SAME_ENTITIES:
+            self.add_error("allocation_entity", _("requirement_and_allocation_entities_cannot_be_different"))
 
     def clean_acronym(self, regex=LEARNING_UNIT_ACRONYM_REGEX_ALL):
         acronym = _merge_first_letter_and_acronym(self.cleaned_data.get('first_letter', ""),
