+++ base/tests/views/test_learning_unit.py
@@ -24,6 +24,7 @@
 #
 ##############################################################################
 import datetime
+import random
 from decimal import Decimal
 from unittest import mock
 
@@ -56,7 +57,8 @@ from base.models.enums import learning_unit_periodicity
 from base.models.enums import learning_unit_year_quadrimesters
 from base.models.enums import learning_unit_year_session
 from base.models.enums import learning_unit_year_subtypes
-from base.models.enums.learning_container_year_types import MASTER_THESIS
+from base.models.enums.learning_container_year_types import MASTER_THESIS, \
+    LEARNING_CONTAINER_YEAR_TYPES_MUST_HAVE_SAME_ENTITIES
 from base.models.enums.learning_unit_year_subtypes import FULL
 from base.models.learning_unit import LearningUnit
 from base.models.learning_unit_year import LearningUnitYear
@@ -141,6 +143,12 @@ class LearningUnitViewTestCase(TestCase):
         self.entity_version = EntityVersionFactory(entity=self.entity, entity_type=entity_type.SCHOOL,
                                                    start_date=today - datetime.timedelta(days=1),
                                                    end_date=today.replace(year=today.year + 1))
+        self.entity_version_2 = EntityVersionFactory(entity=self.entity_2, entity_type=entity_type.INSTITUTE,
+                                                     start_date=today - datetime.timedelta(days=20),
+                                                     end_date=today.replace(year=today.year + 1))
+        self.entity_version_3 = EntityVersionFactory(entity=self.entity_3, entity_type=entity_type.FACULTY,
+                                                     start_date=today - datetime.timedelta(days=50),
+                                                     end_date=today.replace(year=today.year + 1))
 
         self.campus = CampusFactory(organization=self.organization, is_administration=True)
         self.language = LanguageFactory(code='FR')
@@ -750,6 +758,12 @@ class LearningUnitViewTestCase(TestCase):
         faultyDict["acronym"] = ""
         return faultyDict
 
+    def get_faulty_allocation_entity(self):
+        faultyDict = dict(self.get_valid_data())
+        faultyDict["container_type"] = random.choice(LEARNING_CONTAINER_YEAR_TYPES_MUST_HAVE_SAME_ENTITIES)
+        faultyDict["allocation_entity"] = self.entity_version_2.id
+        return faultyDict
+
     def get_faulty_requirement_entity(self):
         """We will create an entity + entity version that user cannot create on it"""
         entity = EntityFactory(country=self.country, organization=self.organization)
@@ -789,6 +803,12 @@ class LearningUnitViewTestCase(TestCase):
         form = CreateLearningUnitYearForm(person=self.person, data=self.get_faulty_acronym())
         self.assertFalse(form.is_valid(), form.errors)
         self.assertEqual(form.errors['acronym'], [_('invalid_acronym')])
+
+    def test_learning_unit_creation_form_with_container_type_same_entity(self):
+        form = CreateLearningUnitYearForm(person=self.person, data=self.get_faulty_allocation_entity())
+        self.assertFalse(form.is_valid(), form.errors)
+        self.assertEqual(form.errors['allocation_entity'],
+                         [_('requirement_and_allocation_entities_cannot_be_different')])
 
     def test_create_learning_unit_case_invalid_academic_year(self):
         now = datetime.datetime.now()
